Restore previous slack status when pomodoro ends
Asynchronously save the user's status_text and status_emoji
when a pomodoro or break is started manually.
Use the saved data to restore the user's status after the end of a pomodoro.
N.B. The restoration is invoked also at the end of every break even
if it is not needed.